Move persistUnreads from sync.ts -> activityActions.ts
